Add transcription job status check to S3-StepFunctions stack
This commit enhances the 's3-stepfunctions-transcription-stack' by adding a check for the transcription job status. It imports additional classes from the 'aws-cdk-lib/aws-stepfunctions' package such as 'Choice', 'Condition', 'Fail', 'Pass', 'Wait', and 'WaitTime'. It defines a 'CheckTranscriptionStatus' task in the StepFunctions stack that invokes the 'getTranscriptionJob' on AWS Transcribe service for checking the status of the transcription job. This is followed by a choice state 'CheckTranscriptionStatusChoice' that makes decisions based on the transcription job status. The commit also includes the timing mechanism to wait for 10 seconds before making a check for the transcription job status. This allows the job some time to complete and ensures that the state machine doesn't fail prematurely.